feat(modal): Pass new client ID to success callback
This change modifies the `CrearClienteModal` to pass the ID of the newly created or updated client to its `onSuccess` callback.

The `CrearTrabajoModal` is updated to receive this new client ID and automatically select the new client in the dropdown list if it's available after the client list refreshes. This improves the user experience by reducing manual steps when creating a new client and then immediately assigning them to a job.